project-4_aws-eks-restapi-deployment-terraform-docker-k8s/app/requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
prometheus-client==0.20.0
python-dotenv